rudimentary support for client and server request info

+++ branches/iiop-net-1-8-0-async_trrefact/IIOPNet/IIOPChannel/PortableInterceptor.cs
@@ -197,7 +197,7 @@ namespace omg.org.PortableInterceptor {
         /// <summary>
         /// request id, which identifies the request/reply sequence.
         /// </summary>
-        long request_id {
+        int request_id {
             get;
         }
